rename to UnscaledValueFunction
