+++ velox/functions/sparksql/UnscaledValueFunction.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "velox/functions/sparksql/DecimalVectorFunctions.h"
+#include "velox/functions/sparksql/UnscaledValueFunction.h"
 
 #include "velox/expression/DecodedArgs.h"
