+++ velox/functions/sparksql/tests/CMakeLists.txt
@@ -37,6 +37,7 @@ add_executable(
   SortArrayTest.cpp
   SplitFunctionsTest.cpp
   StringTest.cpp
+  UnscaledValueFunctionTest.cpp
   XxHash64Test.cpp)
 
 add_test(velox_functions_spark_test velox_functions_spark_test)
